pyproject.toml
[project]
name = "flywheel-fncollector"
version = "0.1.1"
description = "Add a FnCollector for Ryanvk Flywheel"
authors = [
    { name = "F_Qilin", email = "fzq0003@gmail.com" },
]
dependencies = [
    "elaina-flywheel>=0.6.0,<0.7.0",
]
requires-python = ">=3.9"
readme = "README.md"
license = { text = "MIT" }
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Development Status :: 1 - Planning",
]

[project.urls]
Homepage = "https://github.com/FZQ0003/Flywheel-FnCollector"
Issues = "https://github.com/FZQ0003/Flywheel-FnCollector/issues"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm]
distribution = true

[tool.pdm.build]
includes = ["src/flywheel_extras/"]
